+++ tests/models/test_match_spec.py
@@ -10,7 +10,7 @@ from conda import text_type
 from conda.base.context import context
 from conda.cli.common import arg2spec, spec_from_line
 from conda.common.compat import on_win
-from conda.exceptions import CondaValueError
+from conda.exceptions import CondaValueError, InvalidMatchSpec, InvalidSpec
 from conda.models.channel import Channel
 from conda.models.dist import Dist
 from conda.models.records import PackageRecord
@@ -205,6 +205,21 @@ class MatchSpecTests(TestCase):
         assert m("numpy==1.10=py38_0") == "numpy==1.10=py38_0"
         assert m("numpy[version=1.10 build=py38_0]") == "numpy==1.10=py38_0"
 
+        assert m("numpy!=1.10") == "numpy!=1.10"
+        assert m("numpy !=1.10") == "numpy!=1.10"
+        assert m("numpy!=1.10 py38_0") == "numpy[version='!=1.10',build=py38_0]"
+        assert m("numpy !=1.10 py38_0") == "numpy[version='!=1.10',build=py38_0]"
+        assert m("numpy!=1.10=py38_0") == "numpy[version='!=1.10',build=py38_0]"
+        assert m("numpy !=1.10=py38_0") == "numpy[version='!=1.10',build=py38_0]"
+        assert m("numpy >1.7,!=1.10 py38_0") == "numpy[version='>1.7,!=1.10',build=py38_0]"
+        assert m("numpy!=1.10.*") == "numpy!=1.10.*"
+        assert m("numpy!=1.10,!=1.11") == "numpy[version='!=1.10,!=1.11']"
+        assert m("numpy=1.10.*,!=1.10.2") == "numpy[version='=1.10.*,!=1.10.2']"
+
+        assert m("numpy ~=1.10.1") == "numpy~=1.10.1"
+        assert m("numpy~=1.10.1") == "numpy~=1.10.1"
+        assert m("numpy ~=1.10.1 py38_0") == "numpy[version='~=1.10.1',build=py38_0]"
+
         # # a full, exact spec looks like 'defaults/linux-64::numpy==1.8=py26_0'
         # # can we take an old dist str and reliably parse it with MatchSpec?
         # assert m("numpy-1.10-py38_0") == "numpy==1.10=py38_0"
@@ -291,13 +306,13 @@ class MatchSpecTests(TestCase):
         assert str(MatchSpec("defaults::*")) == "defaults::*"
 
     def test_matchspec_errors(self):
-        with pytest.raises(ValueError):
+        with pytest.raises(InvalidSpec):
             MatchSpec('blas [optional')
 
-        with pytest.raises(ValueError):
+        with pytest.raises(InvalidSpec):
             MatchSpec('blas [test=]')
 
-        with pytest.raises(ValueError):
+        with pytest.raises(InvalidSpec):
             MatchSpec('blas[invalid="1"]')
 
         if not on_win:
@@ -425,6 +440,9 @@ class MatchSpecTests(TestCase):
         assert MatchSpec("numpy[build_number='<8']").match(record)
         assert not MatchSpec("numpy[build_number='>7']").match(record)
         assert MatchSpec("numpy[build_number='>=7']").match(record)
+
+        assert MatchSpec("numpy ~=1.10").match(record)
+        assert MatchSpec("numpy~=1.10").match(record)
 
     def test_license_match(self):
         record = {
@@ -648,17 +666,28 @@ class SpecStrParsingTests(TestCase):
             "name": "numpy",
             "version": "1.7*",
         }
+        assert _parse_spec_str("numpy !=1.7") == {
+            "_original_spec_str": "numpy !=1.7",
+            "name": "numpy",
+            "version": "!=1.7",
+        }
 
     def test_parse_hard(self):
+        assert _parse_spec_str("numpy~=1.7.1") == {
+            "_original_spec_str": "numpy~=1.7.1",
+            "name": "numpy",
+            "version": "~=1.7.1",
+        }
         assert _parse_spec_str("numpy>1.8,<2|==1.7") == {
             "_original_spec_str": "numpy>1.8,<2|==1.7",
             "name": "numpy",
             "version": ">1.8,<2|==1.7",
         }
-        assert _parse_spec_str("numpy >1.8,<2|==1.7") == {
-            "_original_spec_str": "numpy >1.8,<2|==1.7",
+        assert _parse_spec_str("numpy >1.8,<2|==1.7,!=1.9,~=1.7.1 py34_0") == {
+            "_original_spec_str": "numpy >1.8,<2|==1.7,!=1.9,~=1.7.1 py34_0",
             "name": "numpy",
-            "version": ">1.8,<2|==1.7",
+            "version": ">1.8,<2|==1.7,!=1.9,~=1.7.1",
+            "build": "py34_0",
         }
         assert _parse_spec_str("*>1.8,<2|==1.7") == {
             "_original_spec_str": "*>1.8,<2|==1.7",
@@ -696,7 +725,7 @@ class SpecStrParsingTests(TestCase):
         }
 
     def test_parse_errors(self):
-        with pytest.raises(CondaValueError):
+        with pytest.raises(InvalidMatchSpec):
             _parse_spec_str('!xyz 1.3')
 
     def test_parse_channel_subdir(self):